rename spotifyClient state slice "app", used for app-wide state
