+++ src/store/app/index.js
@@ -4,10 +4,14 @@ const slice = createSlice({
   name: "spotifyClient",
   initialState: {
     apiToken: "",
+    view: "playlists",
   },
   reducers: {
     setApiToken(state, action) {
       state.apiToken = action.payload.apiToken;
+    },
+    setView(state, action) {
+      state.view = action.payload.view;
     },
   },
 });
