+++ src/store/app/selectors.js
@@ -1,11 +1,5 @@
-export const getPlaylistId = (state) => state.app.playlistId;
-
-export const getLoading = (state) => state.app.loading;
-
-export const getPagination = (state) => state.app.pagination;
-
-export const getSongs = (state) => state.app.songs;
-
-export const getGroups = (state) => state.app.groups;
+export const getApiToken = (state) => state.app.apiToken;
 
 export const getView = (state) => state.app.view;
+
+export const isAuthorised = (state) => state.app.apiToken.length > 0;
