--- src/store/spotify-client/selectors.js
@@ -1,3 +1,0 @@
-export const getApiToken = (state) => state.spotifyClient.apiToken;
-
-export const isAuthorised = (state) => state.spotifyClient.apiToken.length > 0;
