+++ src/store/store.js
@@ -1,15 +1,11 @@
 import { configureStore } from "@reduxjs/toolkit";
 
-import spotifyClient from "./spotify-client";
+import app from "./app";
 import playlists from "./playlists";
-
-// legacy .. maybe unused
-import app from "./app/slice";
 
 const store = configureStore({
   reducer: {
     app,
-    spotifyClient,
     playlists,
   },
 });
